HEEDLS-556 - review markups
