+++ DigitalLearningSolutions.Data.Tests/DataServices/GroupsDataServiceTests.cs
@@ -16,14 +16,12 @@
     {
         private SqlConnection connection = null!;
         private GroupsDataService groupsDataService = null!;
-        private TutorialContentTestHelper tutorialContentTestHelper = null!;
 
         [SetUp]
         public void Setup()
         {
             connection = ServiceTestHelper.GetDatabaseConnection();
             groupsDataService = new GroupsDataService(connection);
-            tutorialContentTestHelper = new TutorialContentTestHelper(connection);
         }
 
         [Test]
@@ -204,145 +202,6 @@
                 // Then
                 groupDelegates.Count.Should().Be(25);
                 groupDelegates.Any(gd => gd.DelegateId == delegateId).Should().BeTrue();
-            }
-            finally
-            {
-                transaction.Dispose();
-            }
-        }
-
-        [Test]
-        public void GetDelegateProgressForCourse_returns_expected_progress()
-        {
-            // Given
-            const int delegateId = 1;
-            const int customisationId = 100;
-
-            // When
-            var delegateProgress = groupsDataService.GetDelegateProgressForCourse(delegateId, customisationId)
-                .FirstOrDefault();
-
-            // Then
-            using (new AssertionScope())
-            {
-                delegateProgress.Should().NotBeNull();
-                delegateProgress?.CandidateId.Should().Be(delegateId);
-                delegateProgress?.CustomisationId.Should().Be(customisationId);
-                delegateProgress?.ProgressId.Should().Be(1);
-                delegateProgress?.Completed.Should().BeNull();
-                delegateProgress?.RemovedDate.Should().BeNull();
-                delegateProgress?.SystemRefreshed.Should().BeFalse();
-                delegateProgress?.SupervisorAdminId.Should().Be(0);
-            }
-        }
-
-        [Test]
-        public void UpdateProgressSupervisorAndCompleteByDate_updates_those_columns()
-        {
-            // Given
-            const int progressId = 1;
-            const int candidateIdForProgressRecord = 1;
-            const int customisationIdForProgressRecord = 100;
-            const int supervisorAdminId = 5;
-            var completeByDate = new DateTime(2021, 12, 25);
-
-            using var transaction = new TransactionScope();
-            try
-            {
-                // When
-                groupsDataService.UpdateProgressSupervisorAndCompleteByDate(
-                    progressId,
-                    supervisorAdminId,
-                    completeByDate
-                );
-                var progressRecords = groupsDataService.GetDelegateProgressForCourse(
-                    candidateIdForProgressRecord,
-                    customisationIdForProgressRecord
-                );
-                var updatedProgressRecord = progressRecords.First(p => p.ProgressId == progressId);
-
-                // Then
-                updatedProgressRecord.SupervisorAdminId.Should().Be(supervisorAdminId);
-                updatedProgressRecord.CompleteByDate.Should().Be(completeByDate);
-            }
-            finally
-            {
-                transaction.Dispose();
-            }
-        }
-
-        [Test]
-        public void CreateNewDelegateProgress_adds_new_record()
-        {
-            // Given
-            const int delegateId = 1;
-            const int customisationId = 100;
-            const int currentVersion = 2;
-            var submittedTime = new DateTime(2021, 3, 3);
-            const int enrollmentMethodId = 3;
-            const int enrolledByAdminId = 5;
-            const int supervisorAdminId = 7;
-            var completeByDate = new DateTime(2020, 1, 1);
-
-            using var transaction = new TransactionScope();
-            try
-            {
-                // When
-                var newProgressId = groupsDataService.CreateNewDelegateProgress(
-                    delegateId,
-                    customisationId,
-                    currentVersion,
-                    submittedTime,
-                    enrollmentMethodId,
-                    enrolledByAdminId,
-                    completeByDate,
-                    supervisorAdminId
-                );
-                var progressRecords = groupsDataService.GetDelegateProgressForCourse(
-                    delegateId,
-                    customisationId
-                );
-                var newProgressRecord = progressRecords.First(p => p.ProgressId == newProgressId);
-
-                // Then
-                using (new AssertionScope())
-                {
-                    newProgressRecord.CandidateId.Should().Be(delegateId);
-                    newProgressRecord.CustomisationId.Should().Be(customisationId);
-                    newProgressRecord.CustomisationVersion.Should().Be(currentVersion);
-                    newProgressRecord.SubmittedTime.Should().Be(submittedTime);
-                    newProgressRecord.EnrollmentMethodId.Should().Be(enrollmentMethodId);
-                    newProgressRecord.EnrolledByAdminId.Should().Be(enrolledByAdminId);
-                    newProgressRecord.SupervisorAdminId.Should().Be(supervisorAdminId);
-                    newProgressRecord.CompleteByDate.Should().Be(completeByDate);
-                    newProgressRecord.Completed.Should().BeNull();
-                    newProgressRecord.RemovedDate.Should().BeNull();
-                    newProgressRecord.SystemRefreshed.Should().BeFalse();
-                }
-            }
-            finally
-            {
-                transaction.Dispose();
-            }
-        }
-
-        [Test]
-        public void CreateNewAspProgressRecord_adds_new_record()
-        {
-            // Given
-            const int progressId = 1;
-            const int tutorialId = 2;
-
-            using var transaction = new TransactionScope();
-            try
-            {
-                // When
-                groupsDataService.CreateNewAspProgress(tutorialId, progressId);
-                var createdAspProgressId =
-                    tutorialContentTestHelper.GetAspProgressFromTutorialAndProgressId(tutorialId, progressId);
-
-                // Then
-                createdAspProgressId.Should().NotBeNull();
             }
             finally
             {
