+++ DigitalLearningSolutions.Data/DataServices/GroupsDataService.cs
@@ -5,7 +5,6 @@
     using System.Data;
     using System.Linq;
     using Dapper;
-    using DigitalLearningSolutions.Data.Models;
     using DigitalLearningSolutions.Data.Models.DelegateGroups;
 
     public interface IGroupsDataService
@@ -24,24 +23,7 @@
 
         void DeleteGroupDelegatesRecordForDelegate(int groupId, int delegateId);
 
-        IEnumerable<Progress> GetDelegateProgressForCourse(int delegateId, int customisationId);
-
-        void UpdateProgressSupervisorAndCompleteByDate(int progressId, int supervisorAdminId, DateTime? completeByDate);
-
         void AddDelegateToGroup(int delegateId, int groupId, DateTime addedDate, int addedByFieldLink);
-
-        int CreateNewDelegateProgress(
-            int delegateId,
-            int customisationId,
-            int customisationVersion,
-            DateTime submittedTime,
-            int enrollmentMethodId,
-            int? enrolledByAdminId,
-            DateTime? completeByDate,
-            int supervisorAdminId
-        );
-
-        void CreateNewAspProgress(int tutorialId, int progressId);
     }
 
     public class GroupsDataService : IGroupsDataService
@@ -204,46 +186,6 @@
             );
         }
 
-        public IEnumerable<Progress> GetDelegateProgressForCourse(
-            int delegateId,
-            int customisationId
-        )
-        {
-            return connection.Query<Progress>(
-                @"SELECT ProgressId, 
-	                    CandidateID, 
-	                    CustomisationID, 
-	                    Completed, 
-	                    RemovedDate, 
-	                    SystemRefreshed, 
-	                    SupervisorAdminID,
-                        CompleteByDate,
-                        EnrollmentMethodID,
-                        EnrolledByAdminID,
-                        SubmittedTime,
-                        CustomisationVersion
-                    FROM Progress 
-                    WHERE CandidateID = @delegateId
-                        AND CustomisationID = @customisationId",
-                new { delegateId, customisationId }
-            );
-        }
-
-        public void UpdateProgressSupervisorAndCompleteByDate(
-            int progressId,
-            int supervisorAdminId,
-            DateTime? completeByDate
-        )
-        {
-            connection.Execute(
-                @"UPDATE Progress SET
-                        SupervisorAdminID = @supervisorAdminId,
-                        CompleteByDate = @completeByDate
-                    WHERE ProgressID = @progressId",
-                new { progressId, supervisorAdminId, completeByDate }
-            );
-        }
-
         public void AddDelegateToGroup(int delegateId, int groupId, DateTime addedDate, int addedByFieldLink)
         {
             connection.Execute(
@@ -254,62 +196,6 @@
                         @addedDate, 
                         @addedByFieldLink)",
                 new { groupId, delegateId, addedDate, addedByFieldLink }
-            );
-        }
-
-        public int CreateNewDelegateProgress(
-            int delegateId,
-            int customisationId,
-            int customisationVersion,
-            DateTime submittedTime,
-            int enrollmentMethodId,
-            int? enrolledByAdminId,
-            DateTime? completeByDate,
-            int supervisorAdminId
-        )
-        {
-            var progressId = connection.QuerySingle<int>(
-                @"INSERT INTO Progress(
-                        CandidateID,
-                        CustomisationID,
-                        CustomisationVersion,
-                        SubmittedTime,
-                        EnrollmentMethodID,
-                        EnrolledByAdminID,
-                        CompleteByDate,
-                        SupervisorAdminID)
-                    OUTPUT Inserted.ProgressID
-                    VALUES (
-                        @delegateId,
-                        @customisationId,
-                        @customisationVersion,
-                        @submittedTime,
-                        @enrollmentMethodId,
-                        @enrolledByAdminId,
-                        @completeByDate,
-                        @supervisorAdminId)",
-                new
-                {
-                    delegateId,
-                    customisationId,
-                    customisationVersion,
-                    submittedTime,
-                    enrollmentMethodId,
-                    enrolledByAdminId,
-                    completeByDate,
-                    supervisorAdminId
-                }
-            );
-
-            return progressId;
-        }
-
-        public void CreateNewAspProgress(int tutorialId, int progressId)
-        {
-            connection.Execute(
-                @"INSERT INTO aspProgress (TutorialId, ProgressId)
-                    VALUES (@tutorialId, @progressId)",
-                new { tutorialId, progressId }
             );
         }
     }
